Se solucionaron los errores al crear la reservacion

+++ reservation-service/src/main/java/com/fastreserve/reservationservice/dto/ReservationResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -18,7 +19,7 @@ public class ReservationResponse {
     private Long userId;
     private Long eventId;
     private Integer seatNumber;
-    private Double totalPrice;
+    private BigDecimal totalPrice;
     private String status;
     private LocalDateTime expirationDate;
 }
